finished chapter name and description

+++ src/controller/chapterControl.ts
@@ -36,6 +36,10 @@ async function getAChapter(req: Request, res: Response) {
 async function deleteChapter(req: Request, res: Response) {
   try {
     const { userId, chapterId } = req.params;
+    if (!userId || !chapterId) {
+      res.status(400).json({ data: "all fields are required" });
+      return;
+    }
     const myChapter = new Chapter(userId, "");
     const response = await myChapter.deleteChapter(chapterId);
     res.status(response.statusCode).json({ data: response.data });
@@ -44,5 +48,47 @@ async function deleteChapter(req: Request, res: Response) {
     res.status(500).json({ data: "internal server Error" });
   }
 }
+async function UpdateChapterName(req: Request, res: Response) {
+  try {
+    const { userId, chapterId } = req.params;
+    const { chapterName } = req.body;
+    if (!userId || !chapterId || !chapterName) {
+      res.status(400).json({ data: "all fields are required" });
+      return;
+    }
+    const myChapter = new Chapter(userId, "");
+    const response = await myChapter.updateChapterName(chapterId, chapterName);
+    res.status(response.statusCode).json({ data: response.data });
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({ data: "internal server Error" });
+  }
+}
+async function UpdateChapterDescription(req: Request, res: Response) {
+  try {
+    const { userId, chapterId } = req.params;
+    const { chapterDescription } = req.body;
+    if (!userId || !chapterId || !chapterDescription) {
+      res.status(400).json({ data: "all fields are required" });
+      return;
+    }
+    const myChapter = new Chapter(userId, "");
+    const response = await myChapter.updateChapterDescription(
+      chapterId,
+      chapterDescription
+    );
+    res.status(response.statusCode).json({ data: response.data });
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({ data: "internal server Error" });
+  }
+}
 
-export default { addNewChapter, getChapters, getAChapter, deleteChapter };
+export default {
+  addNewChapter,
+  getChapters,
+  getAChapter,
+  deleteChapter,
+  UpdateChapterName,
+  UpdateChapterDescription,
+};
